+++ xwiki-rendering-macros/xwiki-rendering-macro-figure/src/main/java/org/xwiki/rendering/internal/macro/figure/FigureMacro.java
@@ -19,30 +19,13 @@
  */
 package org.xwiki.rendering.internal.macro.figure;
 
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-
-import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Singleton;
 
 import org.xwiki.component.annotation.Component;
-import org.xwiki.rendering.block.Block;
-import org.xwiki.rendering.block.CompositeBlock;
 import org.xwiki.rendering.block.FigureBlock;
-import org.xwiki.rendering.block.MacroBlock;
-import org.xwiki.rendering.block.MetaDataBlock;
-import org.xwiki.rendering.block.XDOM;
-import org.xwiki.rendering.macro.AbstractMacro;
-import org.xwiki.rendering.macro.MacroContentParser;
-import org.xwiki.rendering.macro.MacroExecutionException;
-import org.xwiki.rendering.macro.descriptor.DefaultContentDescriptor;
+import org.xwiki.rendering.macro.figure.AbstractFigureMacro;
 import org.xwiki.rendering.macro.figure.FigureMacroParameters;
-import org.xwiki.rendering.macro.figure.FigureType;
-import org.xwiki.rendering.transformation.MacroTransformationContext;
-
-import static org.xwiki.rendering.internal.macro.figure.FigureTypeRecognizerMacro.DATA_XWIKI_RENDERING_FIGURE_TYPE;
 
 /**
  * Tag content as an illustration and with an optional caption.
@@ -53,63 +36,17 @@ import static org.xwiki.rendering.internal.macro.figure.FigureTypeRecognizerMacr
 @Component
 @Named("figure")
 @Singleton
-public class FigureMacro extends AbstractMacro<FigureMacroParameters>
+public class FigureMacro extends AbstractFigureMacro<FigureMacroParameters>
 {
-    /**
-     * The description of the macro.
-     */
-    private static final String DESCRIPTION = "Tag content as an illustration and with an optional caption.";
-
-    /**
-     * The description of the macro content.
-     */
-    private static final String CONTENT_DESCRIPTION = "Illustration(s) and caption";
-
-    @Inject
-    private MacroContentParser contentParser;
-
     /**
      * Create and initialize the descriptor of the macro.
      */
     public FigureMacro()
     {
-        super("Figure", DESCRIPTION, new DefaultContentDescriptor(CONTENT_DESCRIPTION, false, Block.LIST_BLOCK_TYPE),
-            FigureMacroParameters.class);
-        setDefaultCategories(Set.of(DEFAULT_CATEGORY_DEVELOPMENT));
+        super();
     }
 
-    @Override
-    public boolean supportsInlineMode()
-    {
-        return false;
-    }
-
-    @Override
-    public List<Block> execute(FigureMacroParameters parameters, String content, MacroTransformationContext context)
-        throws MacroExecutionException
-    {
-        XDOM xdom = this.contentParser.parse(content, context, false, false);
-        // Mark the macro content as being content that has not been transformed (so that it can be edited inline).
-        List<Block> contentBlock = List.of(new MetaDataBlock(xdom.getChildren(), getNonGeneratedContentMetaData()));
-
-        FigureBlock figureBlock = new FigureBlock(contentBlock);
-        Block block;
-        // If a type is explicitly defined, it is used directly. Otherwise, we try to infer it from the macro's content.
-        if (FigureType.AUTOMATIC != parameters.getType()) {
-            figureBlock.setParameter(DATA_XWIKI_RENDERING_FIGURE_TYPE, parameters.getType().getName());
-            block = figureBlock;
-        } else {
-            block = new CompositeBlock(List.of(
-                new MacroBlock("figureTypeRecognizer", Map.of(), false),
-                figureBlock)
-            );
-        }
-
-        updateStyleParameters(parameters, figureBlock);
-
-        return List.of(block);
-    }
-
+    // TODO: move to platform
     private void updateStyleParameters(FigureMacroParameters parameters, FigureBlock figureBlock)
     {
         if (parameters.getStyle() != null) {
